@@ -3,7 +3,7 @@ const AWS = require('aws-sdk');
 AWS.config.update({ region: 'us-east-1' });
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
 
-const exchangeRatesTableGet = params => (
+const exchangeRatesTableGet = params =>
   new Promise((resolve, reject) => {
     dynamoDb.scan(params, (error, data) => {
       if (error) {
@@ -11,7 +11,6 @@ const exchangeRatesTableGet = params => (
       }
       return resolve(data);
     });
-  })
-);
+  });
 
 module.exports = exchangeRatesTableGet;
