@@ -3,7 +3,7 @@ const AWS = require('aws-sdk');
 AWS.config.update({ region: 'us-east-1' });
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
 
-const exchangeRatesTablePut = params => (
+const exchangeRatesTablePut = params =>
   new Promise((resolve, reject) => {
     dynamoDb.put(params, (error, data) => {
       if (error) {
@@ -11,7 +11,6 @@ const exchangeRatesTablePut = params => (
       }
       return resolve(data);
     });
-  })
-);
+  });
 
 module.exports = exchangeRatesTablePut;
