@@ -1,13 +1,20 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import InputUnit from './InputUnit';
 import ConvertArrows from './ConvertArrows';
 
 const InputUnitGroup = ({
-  leftInput, rightInput, leftUnit, rightUnit, options, handleChange, conversionType, disableInputs,
+  leftInput,
+  rightInput,
+  leftUnit,
+  rightUnit,
+  options,
+  handleChange,
+  conversionType,
+  disableInputs,
 }) => (
-  <Fragment>
+  <>
     <InputUnit
       name="left"
       textValue={leftInput}
@@ -27,7 +34,7 @@ const InputUnitGroup = ({
       conversionType={conversionType}
       disableInputs={disableInputs}
     />
-  </Fragment>
+  </>
 );
 
 InputUnitGroup.defaultProps = {
