@@ -2,49 +2,64 @@ const dataStoragePrefixes = [
   {
     name: 'Kilo',
     abbreviation: 'k',
-  }, {
+  },
+  {
     name: 'Kibi',
     abbreviation: 'Ki',
-  }, {
+  },
+  {
     name: 'Mega',
     abbreviation: 'M',
-  }, {
+  },
+  {
     name: 'Mebi',
     abbreviation: 'Mi',
-  }, {
+  },
+  {
     name: 'Giga',
     abbreviation: 'G',
-  }, {
+  },
+  {
     name: 'Gibi',
     abbreviation: 'Gi',
-  }, {
+  },
+  {
     name: 'Tera',
     abbreviation: 'T',
-  }, {
+  },
+  {
     name: 'Tebi',
     abbreviation: 'Ti',
-  }, {
+  },
+  {
     name: 'Peta',
     abbreviation: 'P',
-  }, {
+  },
+  {
     name: 'Pebi',
     abbreviation: 'Pi',
-  }, {
+  },
+  {
     name: 'Exa',
     abbreviation: 'E',
-  }, {
+  },
+  {
     name: 'Exi',
     abbreviation: 'Ei',
-  }, {
+  },
+  {
     name: 'Zetta',
     abbreviation: 'Z',
-  }, {
+  },
+  {
     name: 'Zebi',
     abbreviation: 'Zi',
-  }, {
+  },
+  {
     name: 'Yotta',
     abbreviation: 'Y',
-  }, {
+  },
+  {
     name: 'Yobi',
     abbreviation: 'Yi',
   },
