@@ -2,31 +2,40 @@ const largePrefixes = [
   {
     name: 'Deca',
     abbreviation: 'da',
-  }, {
+  },
+  {
     name: 'Hecto',
     abbreviation: 'h',
-  }, {
+  },
+  {
     name: 'Kilo',
     abbreviation: 'k',
-  }, {
+  },
+  {
     name: 'Mega',
     abbreviation: 'M',
-  }, {
+  },
+  {
     name: 'Giga',
     abbreviation: 'G',
-  }, {
+  },
+  {
     name: 'Tera',
     abbreviation: 'T',
-  }, {
+  },
+  {
     name: 'Peta',
     abbreviation: 'P',
-  }, {
+  },
+  {
     name: 'Exa',
     abbreviation: 'E',
-  }, {
+  },
+  {
     name: 'Zetta',
     abbreviation: 'Z',
-  }, {
+  },
+  {
     name: 'Yotta',
     abbreviation: 'Y',
   },
