@@ -2,61 +2,80 @@ const metricPrefixes = [
   {
     name: 'Deci',
     abbreviation: 'd',
-  }, {
+  },
+  {
     name: 'Centi',
     abbreviation: 'c',
-  }, {
+  },
+  {
     name: 'Milli',
     abbreviation: 'm',
-  }, {
+  },
+  {
     name: 'Micro',
     abbreviation: 'u',
-  }, {
+  },
+  {
     name: 'Nano',
     abbreviation: 'n',
-  }, {
+  },
+  {
     name: 'Pico',
     abbreviation: 'p',
-  }, {
+  },
+  {
     name: 'Femto',
     abbreviation: 'f',
-  }, {
+  },
+  {
     name: 'Atto',
     abbreviation: 'a',
-  }, {
+  },
+  {
     name: 'Zepto',
     abbreviation: 'z',
-  }, {
+  },
+  {
     name: 'Yocto',
     abbreviation: 'y',
-  }, {
+  },
+  {
     name: 'Deca',
     abbreviation: 'da',
-  }, {
+  },
+  {
     name: 'Hecto',
     abbreviation: 'h',
-  }, {
+  },
+  {
     name: 'Kilo',
     abbreviation: 'k',
-  }, {
+  },
+  {
     name: 'Mega',
     abbreviation: 'M',
-  }, {
+  },
+  {
     name: 'Giga',
     abbreviation: 'G',
-  }, {
+  },
+  {
     name: 'Tera',
     abbreviation: 'T',
-  }, {
+  },
+  {
     name: 'Peta',
     abbreviation: 'P',
-  }, {
+  },
+  {
     name: 'Exa',
     abbreviation: 'E',
-  }, {
+  },
+  {
     name: 'Zetta',
     abbreviation: 'Z',
-  }, {
+  },
+  {
     name: 'Yotta',
     abbreviation: 'Y',
   },
