@@ -2,31 +2,40 @@ const smallPrefixes = [
   {
     name: 'Deci',
     abbreviation: 'd',
-  }, {
+  },
+  {
     name: 'Centi',
     abbreviation: 'c',
-  }, {
+  },
+  {
     name: 'Milli',
     abbreviation: 'm',
-  }, {
+  },
+  {
     name: 'Micro',
     abbreviation: 'u',
-  }, {
+  },
+  {
     name: 'Nano',
     abbreviation: 'n',
-  }, {
+  },
+  {
     name: 'Pico',
     abbreviation: 'p',
-  }, {
+  },
+  {
     name: 'Femto',
     abbreviation: 'f',
-  }, {
+  },
+  {
     name: 'Atto',
     abbreviation: 'a',
-  }, {
+  },
+  {
     name: 'Zepto',
     abbreviation: 'z',
-  }, {
+  },
+  {
     name: 'Yocto',
     abbreviation: 'y',
   },
