@@ -1,6 +1,6 @@
 const addPrefixes = (prefixes, baseUnit) => {
   const units = [];
-  baseUnit.units.forEach((unit) => {
+  baseUnit.units.forEach(unit => {
     units.push({ displayName: unit.displayName, mathName: unit.mathName });
     if (unit.allowPrefixes || baseUnit.mathName === 'dataStorage') {
       const prefixUnits = prefixes.map(prefix => ({
