@@ -3,9 +3,11 @@ import baseUnits from '../unitConstants/baseUnits';
 const getUnits = (inputUnits, conversionType) => {
   const units = {};
   units.leftUnit = inputUnits.leftUnit
-    ? inputUnits.leftUnit : baseUnits[conversionType].initialUnitLeft;
+    ? inputUnits.leftUnit
+    : baseUnits[conversionType].initialUnitLeft;
   units.rightUnit = inputUnits.rightUnit
-    ? inputUnits.rightUnit : baseUnits[conversionType].initialUnitRight;
+    ? inputUnits.rightUnit
+    : baseUnits[conversionType].initialUnitRight;
   return units;
 };
 
